rename dataobjs.pesterQuirks to quirks.PesterQuirkCollection, clean up
